requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
pyinstaller>=6.0.0
